Added Like Button to site.
Included the Facebook JS SDK so that the site can use the XFBML version of Social Plugins. It's a more flexible implementation. I've also placed a Like Button in the left hand sidebar under the existing content. It seemed the most logical place to put it. I used the box_count layout as the standard layout requires 450px width and also doesn't look great when set against a dark background.

Signed-off-by: Colm Doyle <colmdoyle@facebook.com>